pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crypto-predictor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["requests", "PyYAML", "pytest"]

[tool.setuptools.packages.find]
where = [""]          # look in repo root
include = ["src*"]    # package "src" and subpackages
